input "Add external IP check before Ping and MTR"
This commit introduces a function 'IsExternal' in common.go that verifies if a given host is an external IP or not. This function is used in utils.go before performing Ping and MTR to ensure commands are not executed for non-external IPs. In addition, the count display text for the ping result template has been updated.