Add Docker usage instructions to README.md
The Docker usage section was added to README.md file. This was done to provide clear instructions on how to pull the Docker image and start the server and the client, making it easier and more convenient for the users to set up and run the project. As such, user's proficiency level is mitigated as a hindrance to consume our application.